Add files via upload
I didn't like the random channel hopping so I made it sequential.  I also added the 5 Ghz (802.11 a band) channels for adapters like the AWUS051NH that support it.

Depending on the 5 Ghz channel selected, there may be errors from the iw command such as "command failed: Invalid argument (-22)", but most channels will work for capturing packets in the a band.